23/07/22
I finished making the timer page and cleaned up the code a bit.
From my research online and seeking out different ways to create a timer in python. I found a got method and used the concept as a base to work off to create my timer page in my app. I also learnt how to use play sound in python to play an alarm sound once the timer hits 0, but I made for if you exit the timer early sound does not play.